avoid non-integer loop conditions.

+++ gtk2_ardour/automation_controller.cc
@@ -303,10 +303,10 @@ AutomationController::on_button_release(GdkEventButton* ev)
 			                         sigc::mem_fun(*this, &AutomationController::run_note_select_dialog)));
 		}
 		if (is_low) {
-			for (double beats = 1.0; beats <= 16; ++beats) {
+			for (int beats = 1; beats <= 16; ++beats) {
 				items.push_back(MenuElem(string_compose(_("Set to %1 beat(s)"), (int)beats),
 				                         sigc::bind(sigc::mem_fun(*this, &AutomationController::set_freq_beats),
-				                                    beats)));
+				                                    (double)beats)));
 			}
 		}
 		menu->popup(1, ev->time);
